[fusequery/processors] move processors and transforms to pipelines
