+++ fusequery/query/src/interpreters/interpreter_explain.rs
@@ -14,7 +14,7 @@ use log::debug;
 
 use crate::interpreters::IInterpreter;
 use crate::optimizers::Optimizer;
-use crate::processors::PipelineBuilder;
+use crate::pipelines::processors::PipelineBuilder;
 use crate::sessions::FuseQueryContextRef;
 
 pub struct ExplainInterpreter {
