+++ fusequery/query/src/pipelines/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2020-2021 The Datafuse Authors.
+//
+// SPDX-License-Identifier: Apache-2.0.
+
+pub mod processors;
+pub mod transforms;
