+++ fusequery/query/src/pipelines/processors/pipeline_walker_test.rs
@@ -6,7 +6,7 @@
 async fn test_pipeline_walker() -> anyhow::Result<()> {
     use pretty_assertions::assert_eq;
 
-    use crate::processors::*;
+    use crate::pipelines::processors::*;
     use crate::sql::*;
 
     let ctx = crate::tests::try_create_context()?;
