+++ fusequery/query/src/pipelines/transforms/transform_aggregator_final.rs
@@ -16,7 +16,7 @@ use common_streams::{DataBlockStream, SendableDataBlockStream};
 use futures::stream::StreamExt;
 use log::info;
 
-use crate::processors::{EmptyProcessor, IProcessor};
+use crate::pipelines::processors::{EmptyProcessor, IProcessor};
 
 pub struct AggregatorFinalTransform {
     funcs: Vec<Box<dyn IFunction>>,
