+++ fusequery/query/src/pipelines/transforms/transform_limit.rs
@@ -9,7 +9,7 @@ use anyhow::Result;
 use async_trait::async_trait;
 use common_streams::{LimitStream, SendableDataBlockStream};
 
-use crate::processors::{EmptyProcessor, IProcessor};
+use crate::pipelines::processors::{EmptyProcessor, IProcessor};
 
 pub struct LimitTransform {
     limit: usize,
