+++ fusequery/query/src/pipelines/transforms/transform_projection.rs
@@ -13,7 +13,7 @@ use common_functions::IFunction;
 use common_planners::ExpressionPlan;
 use common_streams::{ExpressionStream, SendableDataBlockStream};
 
-use crate::processors::{EmptyProcessor, IProcessor};
+use crate::pipelines::processors::{EmptyProcessor, IProcessor};
 
 pub struct ProjectionTransform {
     funcs: Vec<Box<dyn IFunction>>,
