+++ fusequery/query/src/pipelines/transforms/transform_source_test.rs
@@ -3,36 +3,30 @@
 // SPDX-License-Identifier: Apache-2.0.
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn test_transform_limit() -> anyhow::Result<()> {
+async fn transform_source_test() -> anyhow::Result<()> {
     use std::sync::Arc;
 
-    use common_planners::*;
     use futures::TryStreamExt;
     use pretty_assertions::assert_eq;
 
-    use crate::processors::*;
-    use crate::transforms::*;
+    use crate::pipelines::processors::*;
 
     let ctx = crate::tests::try_create_context()?;
-    let test_source = crate::tests::NumberTestData::create(ctx.clone());
+    let test_source = crate::tests::NumberTestData::create(ctx);
 
     let mut pipeline = Pipeline::create();
 
     let a = test_source.number_source_transform_for_test(8)?;
     pipeline.add_source(Arc::new(a))?;
 
-    pipeline.merge_processor()?;
+    let b = test_source.number_source_transform_for_test(8)?;
 
-    if let PlanNode::Limit(plan) = PlanBuilder::create(test_source.number_schema_for_test()?)
-        .limit(2)?
-        .build()?
-    {
-        pipeline.add_simple_transform(|| Ok(Box::new(LimitTransform::try_create(plan.n)?)))?;
-    }
+    pipeline.add_source(Arc::new(b))?;
+    pipeline.merge_processor()?;
 
     let stream = pipeline.execute().await?;
     let blocks = stream.try_collect::<Vec<_>>().await?;
     let rows: usize = blocks.iter().map(|block| block.num_rows()).sum();
-    assert_eq!(2, rows);
+    assert_eq!(16, rows);
     Ok(())
 }
